Make holiday module generate holidays as needed
This way, it is not required to supply holiday predicates anymore
and you also don't have to generate holiday lists for all years you
might need to have with any given database anymore.

This is a lot cleaner all in all.

The code uses a cache, so lookups are quick the second time a day
for a given year is tested.

Signed-off-by: Frank Terbeck <ft@bewatermyfriend.org>