calendar: Allow precise specification of hours per day
Before, hours-per-day was a fixed number. The problem with this is, when people
start to modify the terms of their contracts within a fiscal year. To handle
this, it is required for hours-per-day to be evaluated for each day. This
allows for that. When hours-per-day is a number, everything works as it did
before. When it is a unary function, it is applied to every day of the
calendar, and that function must return the required number of hours for that
day. Note that the number produced only applies to workdays and compensatory
days, since all other types of days require no work at all.

Signed-off-by: Frank Terbeck <ft@bewatermyfriend.org>